next version of StringReverse. moved to String package. rearranged main. new static metod added.
