--- src/checking/ShowProperties.java
@@ -1,13 +1,0 @@
-package checking;
-
-/**
- * Created by Ivan_Stepantsov on 7/12/2017.
- */
-public class ShowProperties {
-    public static void main(String[] args) {
-        System.getProperties().list(System.out);
-        System.out.println(System.getProperty("user.name"));
-        System.out.println(
-                System.getProperty("java.library.path"));
-    }
-} ///:~
